[esbmc] split BMC strategy driver into parseoptions/
Moves do_bmc_strategy and do_bmc, plus the file-local static
report_non_termination_cwe, into src/esbmc/parseoptions/bmc_strategy.cpp.
do_bmc carries the HAVE_SENDFILE_ESBMC memstats block, so the new file
keeps the guarded sys/sendfile.h include from the original preamble.

Verbatim move. termination, k-induction and k-induction-parallel
regression suites pass unchanged.
