+++ src/esbmc/CMakeLists.txt
@@ -43,6 +43,7 @@ add_custom_command(OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/buildidobj.c
 )
 
 add_executable(esbmc main.cpp esbmc_parseoptions.cpp
+    parseoptions/bmc_strategy.cpp
     parseoptions/function_contracts.cpp
     parseoptions/goto_program.cpp
     parseoptions/process_goto_program.cpp
